# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Queue & Appointment API for Algerian Public Services
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface Wilaya {
  id: number;
  code: string;
  nameAr: string;
  nameFr: string;
  region: string;
}

export interface Commune {
  id: number;
  wilayaId: number;
  nameAr: string;
  nameFr: string;
}

export interface ServiceCategory {
  id: number;
  nameAr: string;
  nameFr: string;
  nameEn: string;
  icon: string;
  description: string;
  color: string;
}

export interface Office {
  id: number;
  name: string;
  nameAr: string;
  categoryId: number;
  categoryName: string;
  wilayaId: number;
  wilayaName: string;
  address: string;
  phone: string;
  openTime: string;
  closeTime: string;
  workingDays: string[];
  averageWaitMinutes: number;
  todayQueueCount: number;
  rating: number;
}

export interface OfficeService {
  id: number;
  officeId: number;
  nameAr: string;
  nameFr: string;
  /** Duration in minutes */
  duration: number;
  requiredDocuments: string[];
  fee: number;
}

export type OfficeDetail = Office & {
  commune: string;
  description: string;
  services: OfficeService[];
};

export interface TimeSlot {
  time: string;
  available: boolean;
  remainingCapacity: number;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  called: "called",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
  no_show: "no_show",
} as const;

export interface Appointment {
  id: number;
  ticketNumber: string;
  officeId: number;
  officeName: string;
  serviceId: number;
  serviceName: string;
  citizenName: string;
  citizenPhone: string;
  citizenNationalId: string;
  date: string;
  time: string;
  status: AppointmentStatus;
  queuePosition: number;
  estimatedWaitMinutes: number;
  createdAt: string;
}

export interface CreateAppointmentRequest {
  officeId: number;
  serviceId: number;
  citizenName: string;
  citizenPhone: string;
  citizenNationalId: string;
  date: string;
  time: string;
  notes?: string;
}

export interface QueueStatus {
  officeId: number;
  officeName: string;
  currentTicket: string;
  currentNumber: number;
  waitingCount: number;
  completedToday: number;
  averageWaitMinutes: number;
  isOpen: boolean;
  lastUpdated: string;
}

export interface QueuePosition {
  appointmentId: number;
  ticketNumber: string;
  position: number;
  estimatedWaitMinutes: number;
  status: string;
}

export type AppointmentAdmin = Appointment & {
  calledAt?: string;
  completedAt?: string;
};

export type UpdateStatusRequestStatus =
  (typeof UpdateStatusRequestStatus)[keyof typeof UpdateStatusRequestStatus];

export const UpdateStatusRequestStatus = {
  confirmed: "confirmed",
  called: "called",
  in_progress: "in_progress",
  completed: "completed",
  no_show: "no_show",
  cancelled: "cancelled",
} as const;

export interface UpdateStatusRequest {
  status: UpdateStatusRequestStatus;
}

export type AdminStatsAppointmentsByHourItem = {
  hour: string;
  count: number;
};

export type AdminStatsAppointmentsByServiceItem = {
  serviceName: string;
  count: number;
};

export interface AdminStats {
  totalToday: number;
  completedToday: number;
  waitingNow: number;
  noShowToday: number;
  averageWaitMinutes: number;
  peakHour: string;
  appointmentsByHour: AdminStatsAppointmentsByHourItem[];
  appointmentsByService: AdminStatsAppointmentsByServiceItem[];
}

export type ListOfficesParams = {
  wilayaId?: number;
  categoryId?: number;
  search?: string;
};

export type GetAvailableSlotsParams = {
  date: string;
  serviceId?: number;
};

export type ListAppointmentsParams = {
  phone: string;
};

export type GetMyQueuePositionParams = {
  appointmentId: number;
};

export type AdminGetQueueParams = {
  date?: string;
};

export type AdminGetStatsParams = {
  officeId?: number;
  date?: string;
};