# lib/api-zod/src/generated/types/appointmentAdmin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Queue & Appointment API for Algerian Public Services
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from "./appointment";

export type AppointmentAdmin = Appointment & {
  calledAt?: Date;
  completedAt?: Date;
};