# lib/api-zod/src/generated/types/queuePosition.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Queue & Appointment API for Algerian Public Services
 * OpenAPI spec version: 0.1.0
 */

export interface QueuePosition {
  appointmentId: number;
  ticketNumber: string;
  position: number;
  estimatedWaitMinutes: number;
  status: string;
}