# lib/api-zod/src/generated/types/queueStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Queue & Appointment API for Algerian Public Services
 * OpenAPI spec version: 0.1.0
 */

export interface QueueStatus {
  officeId: number;
  officeName: string;
  currentTicket: string;
  currentNumber: number;
  waitingCount: number;
  completedToday: number;
  averageWaitMinutes: number;
  isOpen: boolean;
  lastUpdated: Date;
}